test + BoardVO  수정

+++ g_board/src/main/java/com/mycompany/myapp/mapper/BoardMapper.java
@@ -5,5 +5,5 @@ import com.mycompany.myapp.domain.BoardVO;
 public interface BoardMapper {
 	public int count();
 	
-	public BoardVO get(double bNo);
+	public BoardVO get(long bNo);
 }
